backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
beanie==1.23.8
pydantic==2.5.0
python-dotenv==1.0.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4